refactor(#5031): W3 keep test mod in place, widen moved symbols to pub(in discord) (coverage-gate constraint)
